ipipe: drivers/irqchip/irq-vic.c: Fix problems with ipipe in VIC
These changes are needed to enable the VIC interrupt handler for I-pipe usage.
This patch version is for the "ipipe-3.18" branch in Philippes "ipipe" git
respository. But a quick search reveals that such a fix very likely is needed
for kernel version from v3.10 onward.

Tested on SPEAr600 based board with v3.18.

Signed-off-by: Stefan Roese <sr@denx.de>