dovetail: WIP: integrated interface for running co-kernels symbiotically
This work is about implementing the logic for efficient coupling of a
co-kernel in a better way, with deeper integration and conformance
with the existing kernel features and mechanisms.

Interrupt pipelining, timer control and task/context sharing between
the host and co- kernels are in the process of being significantly
reworked. Ultimately, the changes will be reorganized as feature sets.

This is work in progress, and some parts of this code are still in
massive state of flux, including the interface with the dovetailed
code. You should not base anything on it yet, or expect frantic code
churn.