time: account for head IRQ stage preemption in update_process_times()
As the register frame of a timer interrupt preempting the head stage
may be passed to update_process_times(), we should not charge the CPU
time the latter spent to the root activity/process it has preempted.

The typical sequence of events is as follows:

(root stage running)
      [pipelined IRQ, head stage resumed]
(head stage running)
      [pipelined timer IRQ]
      ...
      [pipelined timer IRQ]
      ...
      [head stage suspending]
(root stage running)
      [root timer handling -> update_process_times()]

In that particular sequence, we do not want to charge the timer ticks
received over the head stage to the suspended root context.

To ensure this, we first have to tag the register frame when the
interrupt has preempted the head stage; arch-specific code is usually
clearing the interrupt-enable bit in the CPU state backup area within
the frame to do so.

Next, update_process_times() is changed to receive the register frame
address, checking for the arch-specific marker to know whether the
incoming tick should be charged to the current root context.